Fixed bugs - Renamed javascript_minimal to javascript - Tests ok on Grizzly now
