+++ src/lqc/generate/web_page/javascript/bootstrap_states.js
@@ -1,3 +1,16 @@
+// Get variable by key name from url query string
+// https://css-tricks.com/snippets/javascript/get-url-variables/
+function getQueryVariable(variable)
+{
+  var query = window.location.search.substring(1);
+  var vars = query.split("&");
+  for (var i=0;i<vars.length;i++) {
+    var pair = vars[i].split("=");
+    if(pair[0] == variable){return pair[1];}
+  }
+  return(false);
+}
+
 // Query strings can load page with particular state
 // state=initial - page before any changes are made (default)
 // state=dirty - page after changes
