+++ src/lqc/generate/web_page/javascript/debugging_tools.js
@@ -133,15 +133,3 @@ function checkForBug() {
 
   return compareDimensions(dimensionsAfterApplication, dimensionsAfterFreshLoad);
 }
-
-// https://css-tricks.com/snippets/javascript/get-url-variables/
-function getQueryVariable(variable)
-{
-  var query = window.location.search.substring(1);
-  var vars = query.split("&");
-  for (var i=0;i<vars.length;i++) {
-    var pair = vars[i].split("=");
-    if(pair[0] == variable){return pair[1];}
-  }
-  return(false);
-}
