+++ src/lqc/generate/web_page/javascript/minimal.js
